+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Change `get_scheme_files` and `SchemeFile::new` type arguments from
+  `&Path` to `impl AsRef<Path>` to allow for more flexibility
+
 ## [0.12.1] - 2024-10-06
 
 ### Fixed