+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.13.0] - 2025-03-24
 
 ### Added
 
@@ -226,6 +226,7 @@
 - `sync` subcommand support to sync with latest Tinted Theming schemes
 - `build` subcommand to trigger theme template build
 
+[0.13.0]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.12.1...v0.13.0
 [0.12.1]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.11.1...v0.12.0
 [0.11.1]: https://github.com/tinted-theming/tinted-builder-rust/compare/v0.11.0...v0.11.1